+++ src/component/ModuleContainer.ts
@@ -1,4 +1,4 @@
-import styled from "@alethio/ui/lib/styled-components";
+import styled from "styled-components";
 
 export const ModuleContainer = styled.div`
     position: relative;